backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyMySQL==1.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0
google-generativeai==0.7.2
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.2.0

